requirements.txt
# Required packages for the modern AI Data Analysis Platform
# Install with: pip install -r requirements.txt

streamlit>=1.28.0
supabase>=1.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
langchain>=0.1.0
langgraph>=0.0.40
google-generativeai>=0.3.0
reportlab>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4

# FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Optional for enhanced features
sqlparse>=0.4.0
altair>=5.0.0
pillow>=10.0.0